# Do not allocate in NextCandidate.