… ahocorasick_test.go