@pgavlin 2023-05-24
https://github.com/petar-dambovaliev/aho-corasick/pull/11